[fix] 정수형으로 변경시 누락된 부분 추가 S11P21A108-309

+++ backend/src/main/java/com/example/donzoom/dto/stock/response/StockWalletResponseDto.java
@@ -10,9 +10,9 @@ public class StockWalletResponseDto {
   private final Long stockWalletId;     // 보유 주식 ID
   private final Long stockId;           // 주식 ID
   private final String stockName;       // 주식 이름
-  private final Float totalInvestedPrice; // 총 매입 금액
+  private final Integer totalInvestedPrice; // 총 매입 금액
   private final Integer amount;         // 보유한 주식 수량
-  private final Float averagePrice;     // 평균 매입 단가
+  private final Integer averagePrice;     // 평균 매입 단가
 
   @Builder
   public StockWalletResponseDto(StockWallet stockWallet) {
